// src/app/models/project/exat_pf.ts
export class ExatpfModel {
    constructor() {
     
    }
       
    year_id: string = "2566";

    file_name: string = "";

    file_preix: string = "";
    file_folder: string = "";
    file_enable: boolean = false

    file_order: number = 1;
    file_count_file: number = 0;
        
    modified_by: string = "";
    modified_date: Date = new Date();
   
  }
  